+++ client/src/main/java/rotp/model/ai/base/AIGovernor.java
@@ -82,7 +82,7 @@ public class AIGovernor implements Base, Governor {
         if ((bases > 0) && col.defense().missileBasesCompletedThisTurn())
             session().addSystemToAllocate(sys, text("MAIN_ALLOCATE_BASES_COMPLETE", name, col.defense().maxBases()));
         if (col.industry().isCompletedThisTurn())
-            session().addSystemToAllocate(sys, text("MAIN_ALLOCATE_MAX_FACTORIES", name, (int)col.industry().maxBuildableFactories()));
+            session().addSystemToAllocate(sys, text("MAIN_ALLOCATE_MAX_FACTORIES", name, col.industry().maxBuildableFactories()));
         if (!cleanupOK)
             session().addSystemToAllocate(sys, text("MAIN_ALLOCATE_ECO_LOCKED_WASTE", name));
         if (col.ecology().populationGrowthCompletedThisTurn())
